Only print debugging output for exceptional cases
I noticed that the logging we recently introduced may be too eager. It
seems that it prints each issue to stderr and also to stdout. With this
change, it only prints to stderr when the analysis fails in an
unexpected way, where we would rather `return` to bail out of analysis
earlier than continue.

I'm wondering if we should also `os.Exit(1)` here to make sure the user
is aware that something has gone wrong.